+++ src/r0/account/deactivate.rs
@@ -2,7 +2,9 @@
 
 use ruma_api::ruma_api;
 
-use super::{AuthenticationData, ThirdPartyIdRemovalStatus};
+use crate::r0::uiaa::{AuthData, UiaaResponse};
+
+use super::ThirdPartyIdRemovalStatus;
 
 ruma_api! {
     metadata {
@@ -17,7 +19,7 @@ ruma_api! {
     request {
         /// Additional authentication information for the user-interactive authentication API.
         #[serde(skip_serializing_if = "Option::is_none")]
-        pub auth: Option<AuthenticationData>,
+        pub auth: Option<AuthData>,
         /// Identity server from which to unbind the user's third party
         /// identifier.
         #[serde(skip_serializing_if = "Option::is_none")]
@@ -29,5 +31,5 @@ ruma_api! {
         pub id_server_unbind_result: ThirdPartyIdRemovalStatus,
     }
 
-    error: crate::Error
+    error: UiaaResponse
 }